+++ TaskCoordinatorTest/SharedTypes/TaskCoordinator/Interface/IMessageReader.cs
@@ -5,7 +5,7 @@ namespace TasksCoordinator.Interface
 {
     public interface IMessageReader
     {
-        int taskId { get; }
+        long taskId { get; }
         Task<MessageReaderResult> ProcessMessage(CancellationToken token);
         bool IsPrimaryReader { get; }
     }
